% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datos.R
\name{AdicionarBotonDetalle}
\alias{AdicionarBotonDetalle}
\title{AdicionarBotonDetalle}
\usage{
AdicionarBotonDetalle(tabla)
}
\arguments{
\item{tabla}{Un data frame o tibble. La tabla a la cual se le añadirá el botón de detalle.}
}
\value{
La misma tabla con una nueva columna `Detalle` que contiene un botón interactivo en HTML.
}
\description{
Añade un botón con un ícono de lupa a una tabla para abrir un detalle interactivo.
}
\details{
La función agrega una columna con un ícono de lupa (`&#128270;`) como botón, que puede ser usado para desplegar más información o detalles de un registro específico de la tabla. La función solo realiza modificaciones si la tabla tiene más de una fila.
}
\examples{
\dontrun{
# Crear una tabla de ejemplo
tabla <- data.frame(ID = 1:3, Nombre = c("Juan", "Ana", "Luis"))

# Adicionar el botón de detalle
tabla_con_boton <- AdicionarBotonDetalle(tabla)
print(tabla_con_boton)
}
}
