% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datos.R
\name{AdicionarBotones}
\alias{AdicionarBotones}
\title{Adiciona botones interactivos a una tabla.}
\usage{
AdicionarBotones(tabla, botones)
}
\arguments{
\item{tabla}{Un `data.frame` o `tibble` al que se le desean agregar los botones.}

\item{botones}{Un vector de caracteres que especifica qué botones agregar.
Las opciones disponibles son: "Detalle", "Contacto", "Editar", "Eliminar", "Duplicar".}
}
\value{
Un `data.frame` o `tibble` con nuevas columnas para los botones seleccionados.
}
\description{
Esta función permite agregar botones interactivos a una tabla, cada uno
representando una acción específica. Los botones se pueden usar para
abrir detalles, registrar contactos, editar, eliminar o duplicar.
}
