% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datos.R
\name{CabecerasGraph}
\alias{CabecerasGraph}
\title{CabecerasGraph}
\usage{
CabecerasGraph()
}
\value{
Objeto de cabeceras para usar con httr.
}
\description{
Construye cabeceras HTTP con el token Bearer para Microsoft Graph.
}
\examples{
# headers <- CabecerasGraph()
}
