% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Outputs.R
\name{CajaIco}
\alias{CajaIco}
\title{Generar una Caja con Ícono y Texto}
\usage{
CajaIco(
  texto,
  icono,
  col_fondo = "#FDFEFE",
  alto = 120,
  col_letra = "#17202A",
  col_icono = "#000000"
)
}
\arguments{
\item{texto}{Un carácter que representa el texto que se mostrará en la caja.}

\item{icono}{Un carácter que representa el nombre del ícono a utilizar (usando la librería de íconos de `shiny` o `font-awesome`).}

\item{col_fondo}{Un carácter que representa el color de fondo de la caja (por defecto: `"#FDFEFE"`).}

\item{alto}{Un número que representa la altura mínima de la caja en píxeles (por defecto: `120`).}

\item{col_letra}{Un carácter que representa el color del texto (por defecto: `"#17202A"`).}

\item{col_icono}{Un carácter que representa el color del ícono (por defecto: `"#000000"`).}
}
\value{
Un objeto HTML que contiene la caja con el ícono y el texto estilizados.
}
\description{
Esta función crea una caja visual que incluye un ícono y un texto. La caja puede personalizarse en cuanto a color de fondo, altura, color de letra e ícono.
}
\examples{
CajaIco("Este es un texto", "info-circle", col_fondo = "#E0E0E0", col_letra = "#333333")
}
