% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Formatos.R
\name{col_num}
\alias{col_num}
\title{Asignar color basado en valor numérico}
\usage{
col_num(x)
}
\arguments{
\item{x}{Un valor numérico o vector de valores numéricos.}
}
\value{
Un carácter que representa el color en formato hexadecimal: `#000000` para valores mayores o iguales a cero y `#943126` para valores negativos.
}
\description{
Esta función asigna un color a un número dependiendo de si es positivo, cero o negativo.
}
\examples{
col_num(5)   # Devuelve "#000000"
col_num(-3)  # Devuelve "#943126"
col_num(c(5, -3, 0))  # Devuelve un vector con colores correspondientes a cada valor
}
