% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ElementosGraficos.R
\name{ColoresGreenBlue}
\alias{ColoresGreenBlue}
\title{Generar Colores Graduales entre Verde y Azul}
\usage{
ColoresGreenBlue(value)
}
\arguments{
\item{value}{Un vector de valores para los cuales se desea generar una paleta de colores.
La cantidad de valores determinará cuántos colores se generarán en la paleta.}
}
\value{
Un vector de colores graduales entre verde y azul, con una longitud igual al número de valores de entrada.
}
\description{
Esta función genera una paleta de colores graduales que van desde el color verde
"forestgreen" hasta el color azul oscuro "royalblue4", según el número de valores que se pasen como argumento.
}
\examples{
# Generar una paleta de colores para 5 valores
ColoresGreenBlue(c("A", "B", "C", "D", "E"))

}
