% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datos.R
\name{ConectarBD}
\alias{ConectarBD}
\title{Conectar a la base de datos MySQL usando variables de entorno}
\usage{
ConectarBD()
}
\value{
Un objeto de conexión DBI (DBIConnection).
}
\description{
Establece una conexión a la base de datos MySQL usando las variables de entorno:
DB_NAME, DB_HOST, DB_PORT, DB_USER, DB_PASSWORD, DB_ENCODING.
}
\details{
Esta función ejecuta `SET NAMES 'utf8'` luego de abrir la conexión.
Asegúrate de que las variables de entorno estén definidas en el entorno donde se ejecuta el paquete.
}
\examples{
\dontrun{
Sys.setenv(DB_NAME = "mi_bd", DB_HOST = "localhost", DB_PORT = "3306",
           DB_USER = "usuario", DB_PASSWORD = "secreto", DB_ENCODING = "latin1")
con <- ConectarBD()
DBI::dbDisconnect(con)
}
}
