% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datos.R
\name{ConsultaSistema}
\alias{ConsultaSistema}
\title{Consulta a una base de datos SQL Server}
\usage{
ConsultaSistema(bd, uid, pwd, query)
}
\arguments{
\item{uid}{El nombre de usuario para la conexión a la base de datos.}

\item{pwd}{La contraseña del usuario para la conexión.}

\item{query}{La consulta SQL que se ejecutará en la base de datos.}

\item{base}{Una cadena que especifica la base de datos a la que conectarse. Puede ser
uno de los siguientes valores: "syscafe", "cafesys", o "estad".
El valor correspondiente se traduce al nombre real de la base de datos.}
}
\value{
Un dataframe con los resultados de la consulta, con los nombres de las columnas limpiados.

Un dataframe con los resultados de la consulta y los nombres de las columnas limpiados.
}
\description{
Esta función realiza una consulta a una base de datos SQL Server utilizando ODBC.
Se conecta a diferentes bases de datos dependiendo del valor del parámetro `base`,
ejecuta una consulta SQL y limpia los nombres de las columnas en el dataframe resultante.
}
