% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datos.R
\name{DescargarArchivoId}
\alias{DescargarArchivoId}
\title{DescargarArchivoId}
\usage{
DescargarArchivoId(archivo_id, usuario)
}
\arguments{
\item{archivo_id}{ID del archivo en OneDrive.}

\item{usuario}{Alias del usuario sin dominio.}
}
\value{
Ruta del archivo temporal creado.
}
\description{
Descarga el contenido de un archivo (por ID) de OneDrive a un archivo temporal y devuelve su ruta.
}
\examples{
# tmp <- DescargarArchivoId("ABC123...", "juan.perez")
}
