% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datos.R
\name{DescargarExcelDesdeOneDrive}
\alias{DescargarExcelDesdeOneDrive}
\title{DescargarExcelDesdeOneDrive}
\usage{
DescargarExcelDesdeOneDrive(usuario, ruta, archivo, nombre_salida)
}
\arguments{
\item{usuario}{Alias del usuario sin dominio.}

\item{ruta}{Ruta dentro de OneDrive.}

\item{archivo}{Nombre del archivo en OneDrive.}

\item{nombre_salida}{Nombre base del archivo de salida (sin extensión).}
}
\value{
TRUE si la descarga fue exitosa, en caso contrario lanza error.
}
\description{
Descarga un archivo Excel desde OneDrive y lo guarda localmente con el nombre indicado.
}
\examples{
# ok <- DescargarExcelDesdeOneDrive("juan.perez", "Carpeta/Reportes", "informe.xlsx", "informe_local")
}
