% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fechas.R
\name{EdadCumplida}
\alias{EdadCumplida}
\title{Calcular Edad en Años Entre Dos Fechas}
\usage{
EdadCumplida(from, to)
}
\arguments{
\item{from}{La fecha de nacimiento (fecha inicial).}

\item{to}{La fecha actual (fecha final).}
}
\value{
La edad en años entre las dos fechas.
}
\description{
Esta función calcula la edad en años completos entre dos fechas, considerando los años bisiestos.
}
\examples{
EdadCumplida(as.Date("1990-05-25"), Sys.Date())  # Devuelve la edad actual
EdadCumplida(as.Date("1985-10-10"), as.Date("2023-01-01"))  # Devuelve 37
}
