% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Texto.R
\name{EsEmail}
\alias{EsEmail}
\title{Verificar si una cadena es un correo electrónico válido}
\usage{
EsEmail(email)
}
\arguments{
\item{email}{Cadena de texto que representa el correo electrónico.}
}
\value{
`TRUE` si la cadena corresponde a un correo electrónico válido,
`FALSE` en caso contrario.
}
\description{
Comprueba si una cadena corresponde a un correo electrónico válido.
}
\examples{
EsEmail("test@example.com")  # TRUE
EsEmail("user.name@domain") # FALSE
EsEmail("@example.com")     # FALSE
}
