% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Texto.R
\name{EsEnteroPositivo}
\alias{EsEnteroPositivo}
\title{Verificar si un valor es un entero positivo}
\usage{
EsEnteroPositivo(s)
}
\arguments{
\item{s}{Cadena de texto a evaluar.}
}
\value{
`TRUE` si la cadena representa un entero positivo (sin ceros a la izquierda),
`FALSE` en caso contrario.
}
\description{
Esta función verifica si una cadena representa un entero positivo.
}
\examples{
EsEnteroPositivo("123") # TRUE
EsEnteroPositivo("001") # FALSE
EsEnteroPositivo("-5")  # FALSE
}
