% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Texto.R
\name{EsNumTelefono}
\alias{EsNumTelefono}
\title{Verificar si una cadena es un número de teléfono válido}
\usage{
EsNumTelefono(tel)
}
\arguments{
\item{tel}{Cadena de texto que representa el número de teléfono.}
}
\value{
`TRUE` si la cadena corresponde a un número de teléfono válido,
`FALSE` en caso contrario.
}
\description{
Evalúa si una cadena corresponde a un número de teléfono válido en un formato
específico: debe comenzar con "3" o "6" y tener exactamente 10 dígitos.
}
\examples{
EsNumTelefono("3001234567") # TRUE
EsNumTelefono("6123456789") # TRUE
EsNumTelefono("1234567890") # FALSE
}
