% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Texto.R
\name{EsVacio}
\alias{EsVacio}
\title{Verifica si un valor es NULL, NA o una cadena vacía}
\usage{
EsVacio(x)
}
\arguments{
\item{x}{Valor o vector de valores a verificar.}
}
\value{
Un valor lógico: `TRUE` si `x` es `NULL` o si algún elemento es `NA` o una cadena vacía
(tras aplicar `trimws`), de lo contrario `FALSE`.
}
\description{
Esta función evalúa un valor y determina si es NULL, NA o una cadena vacía ("").
}
\examples{
EsVacio(NULL)              # TRUE
EsVacio(c(NA, ""))        # TRUE
EsVacio(c("texto", NA))   # TRUE
EsVacio(c(" ", "texto")) # TRUE
EsVacio(c("a", "b"))     # FALSE
}
