% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fechas.R
\name{FechaTexto}
\alias{FechaTexto}
\title{Convertir una Fecha en un Formato de Texto Personalizado}
\usage{
FechaTexto(
  x,
  dia = TRUE,
  dia_nombre = TRUE,
  dia_nom_abr = TRUE,
  mes = TRUE,
  mes_abr = TRUE,
  anho = TRUE,
  anho_abr = TRUE,
  sep_texto = TRUE
)
}
\arguments{
\item{x}{Un vector de fechas que se desea convertir.}

\item{dia}{Un valor lógico que indica si se debe incluir el día (predeterminado: TRUE).}

\item{dia_nombre}{Un valor lógico que indica si se debe incluir el nombre completo del día (predeterminado: TRUE).}

\item{dia_nom_abr}{Un valor lógico que indica si se debe incluir el nombre abreviado del día (predeterminado: TRUE).}

\item{mes}{Un valor lógico que indica si se debe incluir el mes (predeterminado: TRUE).}

\item{mes_abr}{Un valor lógico que indica si se debe incluir el nombre abreviado del mes (predeterminado: TRUE).}

\item{anho}{Un valor lógico que indica si se debe incluir el año (predeterminado: TRUE).}

\item{anho_abr}{Un valor lógico que indica si se debe incluir el año abreviado (predeterminado: TRUE).}

\item{sep_texto}{Un valor lógico que indica si se debe incluir un separador de texto entre las partes de la fecha (predeterminado: TRUE).}
}
\value{
Un vector de cadenas de texto que representa las fechas en el formato especificado.
}
\description{
Esta función convierte una fecha en un formato de texto personalizado, permitiendo seleccionar qué partes de la fecha mostrar (día, mes, año).
}
\examples{
FechaTexto(as.Date("2023-10-15"), dia = TRUE)
FechaTexto(as.Date("2023-10-15"), dia = TRUE, dia_nombre = FALSE)
}
