% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Formatos.R
\name{FormatearNumero}
\alias{FormatearNumero}
\title{Formatear un Número con Opciones de Estilo y Colores}
\usage{
FormatearNumero(
  x,
  formato,
  negrita = TRUE,
  color = "#000000",
  meta = NA,
  prop = TRUE
)
}
\arguments{
\item{x}{El número o vector de números a formatear.}

\item{formato}{El formato deseado para el número. Opciones incluyen:
"coma" (con separadores de miles), "numero", "dinero", "miles", "porcentaje".}

\item{negrita}{Un valor lógico (`TRUE` o `FALSE`) que indica si el número debe aparecer en negrita (por defecto es `TRUE`).}

\item{color}{El color en formato hexadecimal (por defecto es "#000000"). Si no se proporciona `meta`, este color será usado.}

\item{meta}{Un valor o vector numérico que representa el umbral o meta para cambiar el color. (opcional)}

\item{prop}{Un valor lógico (`TRUE` o `FALSE`) que indica si se debe usar una escala de colores proporcional al valor de `x` con respecto a `meta` (por defecto es `TRUE`).}
}
\value{
Un número formateado en HTML con opciones de estilo (negrita, color) y el formato numérico especificado.
}
\description{
Esta función permite formatear un número de acuerdo con un formato específico,
con opciones para aplicar negrita y cambiar el color en función de un umbral
o meta. Se puede utilizar para resaltar visualmente los números en gráficos o tablas.
}
\examples{
# Formatear un número en formato "dinero" con color proporcional
FormatearNumero(2500, "dinero", negrita = TRUE, meta = 2000)

# Formatear un número en formato "porcentaje" sin negrita y con color específico
FormatearNumero(0.75, "porcentaje", negrita = FALSE, color = "#00FF00")

}
