% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Formatos.R
\name{FormatoD3}
\alias{FormatoD3}
\title{Definir formato para D3.js}
\usage{
FormatoD3(formato)
}
\arguments{
\item{formato}{Cadena de texto con el formato deseado. Valores permitidos:
"coma", "numero", "dinero", "dolares", "miles", "porcentaje",
"cientifico", "millones", "entero", "tiempo", "kwh" o "log".}
}
\value{
Un string que representa el formato en D3.js.
}
\description{
Genera un string de formato compatible con la librería D3.js.
}
\examples{
FormatoD3("coma")
FormatoD3("dinero")

}
