% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Formatos.R
\name{gt_minimal_style}
\alias{gt_minimal_style}
\title{Estilo minimalista para tablas gt}
\usage{
gt_minimal_style(gt_table)
}
\arguments{
\item{gt_table}{Un objeto de clase `gt_tbl` creado previamente con el paquete `gt`.}
}
\value{
Un objeto `gt_tbl` con estilos de tabla personalizados.
}
\description{
Aplica un tema visual limpio y minimalista a un objeto de tabla creado con
el paquete **gt**, eliminando casi todos los bordes y resaltando únicamente
los encabezados de columna con una línea sutil.
}
\examples{
library(gt)
library(dplyr)

# Ejemplo básico
df <- head(mtcars)
tab <- gt(df)
gt_minimal_style(tab)

}
