% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTStyles.R
\name{gt_var_style}
\alias{gt_var_style}
\title{Aplicar estilo de variación a columnas de una tabla `gt`}
\usage{
  gt_var_style(gt_table, ...)
}
\arguments{
\item{gt_table}{Un objeto de clase [`gt::gt()`] al que se le aplicarán los estilos.}

\item{...}{Nombres de columnas dentro de `gt_table` a los que se aplicará el estilo.}
}
\value{
El objeto `gt_table` con los estilos agregados.
}
\description{
Colorea el texto de columnas numéricas de una tabla `gt` para distinguir valores
negativos y positivos.
}
\examples{
\dontrun{
library(gt)
data <- data.frame(etiqueta = c("A", "B"), variacion = c(-0.1, 0.25))
data %>%
  gt() %>%
  gt_var_style(variacion)
}
}
\seealso{
Other estilos_gt: \code{\link{gt_color_columns}()}, \code{\link{gt_pct_style}()}
}
