% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ElementosGraficos.R
\name{hline}
\alias{hline}
\title{Crear Línea Horizontal en un Gráfico}
\usage{
hline(y = 0, color = "#ff3a21")
}
\arguments{
\item{y}{Posición en el eje y donde se desea trazar la línea (por defecto: 0).}

\item{color}{Color de la línea (por defecto: "#ff3a21").}
}
\value{
Una lista con las propiedades necesarias para dibujar la línea horizontal en un gráfico.
}
\description{
Esta función genera una línea horizontal que puede ser utilizada en gráficos interactivos,
como los creados con la librería \code{plotly}. La línea se dibuja en una posición especificada
sobre el eje y y se puede personalizar su color.
}
\examples{
hline(3)  # Traza una línea horizontal en y = 3 con color #ff3a21.
hline(5, "green")  # Traza una línea horizontal en y = 5 con color verde.
}
