% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Outputs.R
\name{ImprimeSankey}
\alias{ImprimeSankey}
\title{Imprime un diagrama de Sankey utilizando datos agrupados.}
\usage{
ImprimeSankey(data, vars, fun, var = NULL, colores)
}
\arguments{
\item{data}{Un data frame que contiene los datos a visualizar.}

\item{vars}{Un vector de caracteres que especifica las variables a agrupar.}

\item{fun}{Una función de resumen que se aplicará a `var` (por ejemplo, "n" para contar).}

\item{var}{Una variable numérica opcional que se utilizará con `fun` para calcular el resumen.}

\item{colores}{Un vector de caracteres que define los colores para cada variable en `vars`.}
}
\value{
Una lista que contiene el objeto `plot` del diagrama de Sankey,
        el data frame `nodos` y el data frame `links`.
}
\description{
Esta función crea un diagrama de Sankey que representa las relaciones entre múltiples variables.
Permite resumir los datos utilizando una función específica y personalizar los colores de los nodos.
}
