% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inputs.R
\name{InputNumerico}
\alias{InputNumerico}
\title{Crear un input numérico estilizado para Shiny}
\usage{
InputNumerico(
  id,
  label,
  value,
  dec = 2,
  max = NULL,
  min = NULL,
  type = "numero"
)
}
\arguments{
\item{id}{Identificador único del input.}

\item{label}{Etiqueta que describe el campo.}

\item{value}{Valor inicial del input.}

\item{dec}{Número de decimales a mostrar (por defecto 2).}

\item{max}{Valor máximo permitido (por defecto NULL).}

\item{min}{Valor mínimo permitido (por defecto NULL).}

\item{type}{Tipo de input: "dinero", "porcentaje" o "numero".}
}
\value{
Un objeto de tipo `fluidRow` con el diseño del input.
}
\description{
Crear un input numérico estilizado para Shiny
}
\examples{
# Ejemplo para un input de dinero
InputGenerico("dinero_input", "Monto:", 1000, type = "dinero")

# Ejemplo para un input de porcentaje
InputGenerico("porcentaje_input", "Porcentaje:", 50, type = "porcentaje")

# Ejemplo para un input numérico general
InputGenerico("numero_input", "Cantidad:", 10, max = 100, min = 0, dec = 3, type = "numero")
}
