% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Html.R
\name{labelObligatorio}
\alias{labelObligatorio}
\title{Crear una Etiqueta Obligatoria con un Asterisco}
\usage{
labelObligatorio(label)
}
\arguments{
\item{label}{Texto de la etiqueta que se desea mostrar.}
}
\value{
Un objeto `tagList` que representa la etiqueta HTML con el asterisco visual.
}
\description{
Esta función genera una etiqueta HTML que incluye el texto de la etiqueta y un asterisco (`*`)
para indicar que el campo es obligatorio. El asterisco es presentado con una clase CSS llamada
`mandatory_star` para facilitar el estilo visual.
}
\examples{
# Crear una etiqueta obligatoria
labelObligatorio("Nombre del Usuario")

}
