% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Texto.R
\name{LimpiarNombres}
\alias{LimpiarNombres}
\title{Limpiar y Normalizar Nombres}
\usage{
LimpiarNombres(s)
}
\arguments{
\item{s}{Un vector de caracteres con los nombres o cadenas de texto a limpiar.}
}
\value{
Un vector de caracteres con las cadenas normalizadas, sin espacios en blanco duplicados y en mayúsculas.
}
\description{
Esta función limpia y normaliza una cadena de texto eliminando espacios repetidos y
convirtiendo todo a mayúsculas. Se asegura de que los espacios múltiples sean reducidos a uno solo
y elimina espacios en blanco al principio y al final de la cadena.
}
\examples{
LimpiarNombres("  Camilo    Yate     Támara   ")  # Devuelve "CAMILO YATE TÄMARA"
}
