% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datos.R
\name{ListarCarpetas}
\alias{ListarCarpetas}
\title{ListarCarpetas}
\usage{
ListarCarpetas(usuario)
}
\arguments{
\item{usuario}{Alias del usuario sin dominio.}
}
\value{
Vector de caracteres con nombres de carpetas.
}
\description{
Lista el nombre de las carpetas en la raíz del OneDrive del usuario.
}
\examples{
# carpetas <- ListarCarpetas("juan.perez")
}
