% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datos.R
\name{ListarContenidoCarpetaId}
\alias{ListarContenidoCarpetaId}
\title{ListarContenidoCarpetaId}
\usage{
ListarContenidoCarpetaId(usuario, carpeta_id)
}
\arguments{
\item{usuario}{Alias del usuario sin dominio.}

\item{carpeta_id}{ID de la carpeta.}
}
\value{
Tibble con columnas: name, type, id.
}
\description{
Lista los elementos (archivos y carpetas) dentro de una carpeta (por ID).
}
\examples{
# df <- ListarContenidoCarpetaPorId("juan.perez", "0123ABC...")
}
