% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datos.R
\name{ListarContenidoCarpetaNombre}
\alias{ListarContenidoCarpetaNombre}
\title{ListarContenidoCarpetaNombre}
\usage{
ListarContenidoCarpetaNombre(usuario, nombre_carpeta)
}
\arguments{
\item{usuario}{Alias del usuario sin dominio.}

\item{nombre_carpeta}{Nombre de la carpeta.}
}
\value{
Tibble con columnas: name, type, id.
}
\description{
Lista los elementos (archivos y carpetas) dentro de una carpeta especificada por nombre en la raíz.
}
\examples{
# df <- ListarContenidoCarpetaPorNombre("juan.perez", "Reportes")
}
