% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datos.R
\name{ListarHojasExcelOneDrive}
\alias{ListarHojasExcelOneDrive}
\title{ListarHojasExcelOneDrive}
\usage{
ListarHojasExcelOneDrive(archivo_id, usuario)
}
\arguments{
\item{archivo_id}{ID del archivo en OneDrive.}

\item{usuario}{Alias del usuario sin dominio.}
}
\value{
Vector de caracteres con los nombres de hojas del Excel.
}
\description{
Descarga un archivo Excel (por ID) y retorna el vector de nombres de hojas.
}
\examples{
# tmp <- DescargarArchivoId("ABC123...", "juan.perez")
# hojas <- ListarHojasExcelOneDrive("ABC123...", "juan.perez")
}
