% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datos.R
\name{ListarTodoContenidoCarpeta}
\alias{ListarTodoContenidoCarpeta}
\title{ListarTodoContenidoCarpeta}
\usage{
ListarTodoContenidoCarpeta(usuario, carpeta_id)
}
\arguments{
\item{usuario}{Alias del usuario sin dominio.}

\item{carpeta_id}{ID de la carpeta raíz.}
}
\value{
Tibble con columnas: name, type, id, path.
}
\description{
Retorna un tibble con todos los archivos dentro de una carpeta (por ID) de manera recursiva.
}
\examples{
# df <- ListarTodoContenidoCarpeta("juan.perez", "0123ABC...")
}
