% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Liberias.R
\name{Loadpkg}
\alias{Loadpkg}
\title{Cargar e Instalar Paquetes Requeridos}
\usage{
Loadpkg(pkg)
}
\arguments{
\item{pkg}{Un vector de caracteres con los nombres de los paquetes que se desean verificar, instalar (si es necesario) y cargar.}
}
\value{
Un vector lógico que indica si cada paquete fue cargado con éxito (\code{TRUE} si se cargó, \code{FALSE} en caso contrario).
}
\description{
Esta función verifica si los paquetes de R especificados están instalados.
Si algún paquete no está instalado, lo instala. Posteriormente, carga todos los paquetes solicitados.
}
\examples{
Loadpkg(c("ggplot2", "dplyr"))  # Instala y carga ggplot2 y dplyr si es necesario
}
