% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Numericos.R
\name{Moda}
\alias{Moda}
\title{Calcular la Moda de un Vector}
\usage{
Moda(x, na.rm = TRUE)
}
\arguments{
\item{x}{El vector numérico o categórico del cual se desea calcular la moda.}

\item{na.rm}{Un valor lógico que indica si se deben excluir los valores `NA` en el cálculo de la moda.
El valor predeterminado es `TRUE`.}
}
\value{
El valor de la moda del vector, es decir, el valor que aparece con mayor frecuencia. Si hay más de una moda, devuelve el primero encontrado.
}
\description{
Esta función calcula la moda de un vector numérico o categórico. La moda es el valor que aparece con mayor frecuencia en el vector.
Si hay más de una moda, la función devuelve solo una de ellas.
}
\examples{
Moda(c(1, 2, 2, 3, 4))  # Devuelve 2
Moda(c("a", "b", "b", "c", "c", "c"))  # Devuelve "c"
Moda(c(1, 1, 2, 2, 3), na.rm = TRUE)  # Devuelve 1
}
