% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Html.R
\name{Obligatorio}
\alias{Obligatorio}
\title{Obligatorio: Añade un asterisco rojo a un campo obligatorio}
\usage{
Obligatorio(s)
}
\arguments{
\item{s}{Un string que representa el texto de la etiqueta que será mostrado en el formulario.}
}
\value{
Un objeto `HTML` que será renderizado en la interfaz de usuario de `shiny`, mostrando una etiqueta con un asterisco rojo que denota obligatoriedad.
}
\description{
Esta función genera una etiqueta que indica que un campo es obligatorio, añadiendo un asterisco de color rojo junto al texto de la etiqueta.
}
\details{
La función envuelve el texto de la etiqueta proporcionada en un encabezado HTML de nivel 6 (`h6`) y añade un asterisco en color rojo, utilizando la funcionalidad de `shiny`.
}
\examples{
Obligatorio("Nombre")  # Crea una etiqueta con la palabra "Nombre" y un asterisco rojo.
}
