% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datos.R
\name{ObtenerIdCarpeta}
\alias{ObtenerIdCarpeta}
\title{ObtenerIdCarpeta}
\usage{
ObtenerIdCarpeta(usuario, nombre_carpeta)
}
\arguments{
\item{usuario}{Alias del usuario sin dominio.}

\item{nombre_carpeta}{Nombre exacto de la carpeta.}
}
\value{
Cadena con el ID de la carpeta si existe; en caso contrario, error.
}
\description{
Obtiene el ID de una carpeta en la raíz del OneDrive del usuario por su nombre.
}
\examples{
# carpeta_id <- ObtenerIdCarpeta("juan.perez", "Reportes")
}
