% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datos.R
\name{ObtenerIdDrive}
\alias{ObtenerIdDrive}
\title{ObtenerIdDrive}
\usage{
ObtenerIdDrive(usuario)
}
\arguments{
\item{usuario}{Cadena con el alias o correo (sin dominio) del usuario.}
}
\value{
Cadena con el identificador de la unidad (drive id) del usuario.
}
\description{
Obtiene el ID de OneDrive (drive id) de un usuario de dominio racafe.com.
}
\examples{
# drive_id <- ObtenerIdUnidad("hcyate")
}
