% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inputs.R
\name{pick_opt}
\alias{pick_opt}
\title{Genera opciones para selectPicker}
\usage{
pick_opt(cho, fem = TRUE)
}
\arguments{
\item{cho}{Vector con las opciones del selector. Se utiliza para configurar límites del formato de selección.}

\item{fem}{Lógico. Indica si el texto debe ser femenino ("Todas") o masculino ("Todos"). Por defecto, \code{TRUE} (femenino).}
}
\value{
Una lista con las opciones necesarias para personalizar el comportamiento de `selectPicker`.
}
\description{
Esta función construye una lista de opciones personalizadas para el widget `selectPicker` de la librería `shinyWidgets`.
Permite configurar opciones de selección masiva y formato de texto seleccionado.
}
\examples{
pick_opt(letters) # Opciones con valores femeninos "Todas"
pick_opt(letters, fem = FALSE) # Opciones con valores masculinos "Todos"
}
