% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fechas.R
\name{PrimerDia}
\alias{PrimerDia}
\title{Obtener el Primer Día de una Unidad de Tiempo en una Fecha}
\usage{
PrimerDia(x, uni = "month")
}
\arguments{
\item{x}{La fecha de la cual se desea obtener el primer día. Puede ser un objeto de clase \code{Date} o \code{POSIXt}.}

\item{uni}{La unidad de tiempo que se desea usar para el cálculo. Puede ser "month", "year", entre otras. El valor predeterminado es "month".}
}
\value{
Un objeto de clase \code{Date} que representa el primer día de la unidad temporal de la fecha dada.
}
\description{
Esta función calcula el primer día de la unidad temporal especificada (mes, año, etc.) de una fecha dada.
Si no se especifica, el valor predeterminado es calcular el primer día del mes.
}
\examples{
PrimerDia("2023-10-15")  # Devuelve "2023-10-01"
PrimerDia(as.Date("2023-05-22"))  # Devuelve "2023-05-01"
PrimerDia("2023-10-15", uni = "year")  # Devuelve "2023-01-01"
}
