% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Html.R
\name{Saltos}
\alias{Saltos}
\title{Insertar Saltos de Línea HTML}
\usage{
Saltos(n = 1)
}
\arguments{
\item{n}{Un valor entero que indica el número de saltos de línea a generar. El valor predeterminado es \code{1}.}
}
\value{
Una cadena de texto que contiene \code{<br/>} repetido \code{n} veces.
}
\description{
Esta función genera una cadena con un número especificado de saltos de línea en formato HTML (\code{<br/>}).
}
\examples{
Saltos(3)  # Devuelve "<br/><br/><br/>"
}
