% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/racafeShiny.R
\name{SidebarItemWrap}
\alias{SidebarItemWrap}
\title{Envolver un elemento del sidebar en un contenedor con ID}
\usage{
SidebarItemWrap(label, tabName, icon)
}
\arguments{
\item{label}{Etiqueta visible del elemento del menú.}

\item{tabName}{Nombre de la pestaña asociada al elemento.}

\item{icon}{Ícono del elemento del menú.}
}
\value{
Un contenedor \code{shiny::div()} con ID \code{wrap_\{tabName\}} que incluye
el \code{bs4SidebarMenuItem}.
}
\description{
Esta función envuelve un \code{bs4Dash::bs4SidebarMenuItem()} en un \code{div}
cuyo ID sigue el patrón \code{wrap_\{tabName\}} para facilitar reglas CSS
globales como \code{[id^='wrap_']}.
}
