% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datos.R
\name{TopRelativo}
\alias{TopRelativo}
\title{Recodificación de Categorías Menos Frecuentes (Relativa)}
\usage{
TopRelativo(
  data,
  var_recode,
  var_top,
  fun_Top,
  pct_min = 0.05,
  nom_var,
  lab_recodificar = "OTROS"
)
}
\arguments{
\item{data}{El conjunto de datos en el cual se encuentra la variable a recodificar.}

\item{var_recode}{Variable que se desea recodificar. Se pasa sin comillas.}

\item{var_top}{Variable a partir de la cual se calcularán las frecuencias o la función de resumen. Se pasa sin comillas.}

\item{fun_Top}{La función de resumen a aplicar en caso de no utilizar las frecuencias absolutas (por ejemplo, "mean", "sum", etc.).}

\item{pct_min}{El porcentaje mínimo necesario para considerar una categoría principal (predeterminado: 0.05).}

\item{nom_var}{El nombre para la nueva variable recodificada.}

\item{lab_recodificar}{El nombre o etiqueta para las categorías recodificadas (predeterminado: "OTROS").}
}
\value{
El conjunto de datos con la variable recodificada según las categorías principales y las categorías recodificadas.
}
\description{
Recodifica las categorías menos frecuentes de una variable según su valor relativo o una función de resumen y las agrupa en una nueva categoría.
}
