% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Texto.R
\name{UnirCadenas}
\alias{UnirCadenas}
\title{Unir Múltiples Cadenas de Texto}
\usage{
UnirCadenas(..., sep = " ", collapse = NULL, na.rm = FALSE)
}
\arguments{
\item{...}{Varios vectores de caracteres que se desean unir.}

\item{sep}{Separador entre las cadenas. El valor predeterminado es un espacio en blanco (\code{" "}).}

\item{collapse}{Un valor opcional para definir un separador entre los elementos resultantes,
similar a \code{paste}'s \code{collapse}. El valor predeterminado es \code{NULL}.}

\item{na.rm}{Un valor lógico que indica si los valores \code{NA} deben ser eliminados.
El valor predeterminado es \code{FALSE}, lo que incluye los \code{NA} en el resultado.}
}
\value{
Una cadena de texto resultante de la unión de las cadenas especificadas.
}
\description{
Esta función une múltiples cadenas de texto en una sola cadena, con la opción de omitir valores \code{NA}.
Es similar a \code{paste}, pero permite controlar la eliminación de valores ausentes.
}
\examples{
UnirCadenas("Hola", NA, "Mundo", sep = "-", na.rm = TRUE)
# Devuelve "Hola-Mundo"
UnirCadenas("Hola", NA, "Mundo", sep = " ", na.rm = FALSE)
# Devuelve "Hola NA Mundo"
}
