% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ElementosGraficos.R
\name{vline}
\alias{vline}
\title{Crear Línea Vertical en un Gráfico}
\usage{
vline(x = 0, color = "red")
}
\arguments{
\item{x}{Posición en el eje x donde se desea trazar la línea (por defecto: 0).}

\item{color}{Color de la línea (por defecto: "red").}
}
\value{
Una lista con las propiedades necesarias para dibujar la línea vertical en un gráfico.
}
\description{
Esta función genera una línea vertical que puede ser utilizada en gráficos interactivos,
como los creados con la librería \code{plotly}. La línea se dibuja en una posición especificada
sobre el eje x y se puede personalizar su color.
}
\examples{
vline(2)  # Traza una línea vertical en x = 2 con color rojo.
vline(3, "blue")  # Traza una línea vertical en x = 3 con color azul.
}
