% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Formatos.R
\name{chr_kpi}
\alias{chr_kpi}
\title{Símbolo para KPI}
\usage{
chr_kpi(x)
}
\arguments{
\item{x}{Valor numérico.}
}
\value{
Un carácter: ▲ (positivo), ▼ (negativo), ▬ (cero).
}
\description{
Devuelve un símbolo gráfico que representa la dirección de un indicador.
}
\examples{
chr_kpi(1)
chr_kpi(-1)

}
