% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Formatos.R
\name{col_kpi}
\alias{col_kpi}
\title{Color para KPI}
\usage{
col_kpi(x, prop = TRUE)
}
\arguments{
\item{x}{Valor numérico.}

\item{prop}{Lógico, indica si el KPI es proporcional (default: TRUE).}
}
\value{
Cadena con un color en formato hexadecimal.
}
\description{
Asigna color según si el valor es positivo, negativo o cero, y según proporcionalidad.
}
\examples{
col_kpi(1, TRUE)
col_kpi(-1, FALSE)

}
