% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Formatos.R
\name{col_kpi}
\alias{col_kpi}
\title{Obtener el color para un indicador}
\usage{
col_kpi(x, prop = TRUE)
}
\arguments{
\item{x}{El valor del indicador.}

\item{prop}{Un valor lógico que indica si el indicador es proporcional (TRUE) o no (FALSE).
El valor predeterminado es TRUE.}
}
\value{
El color correspondiente como una cadena hexadecimal.
}
\description{
Esta función devuelve el color correspondiente para un indicador, según si es proporcional o no.
}
\examples{
col_kpi(1, TRUE)  # Indicador proporcional positivo (color verde).
col_kpi(-1, FALSE)  # Indicador no proporcional negativo (color rojo).
}
