% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Formatos.R
\name{col_num}
\alias{col_num}
\title{Color según número}
\usage{
col_num(x)
}
\arguments{
\item{x}{Valor numérico o vector.}
}
\value{
Vector de colores en formato hexadecimal.
}
\description{
Asigna color a un número: negro si es mayor o igual a cero y rojo oscuro si es negativo.
}
\examples{
col_num(5)
col_num(-3)
col_num(c(5, -3, 0))

}
