% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Formatos.R
\name{DefinirFormato}
\alias{DefinirFormato}
\title{Definir Formato para Números}
\usage{
DefinirFormato(formato)
}
\arguments{
\item{formato}{Un string que especifica el formato deseado para los números. Puede ser uno de los siguientes:
- "coma": Números con comas como separadores de miles.
- "numero": Números con dos decimales.
- "dinero": Números con símbolo "$" y comas como separadores de miles.
- "miles": Números en miles con símbolo "$" y comas como separadores de miles.
- "porcentaje": Números expresados como porcentaje (multiplicados por 100 y con "%" al final).}
}
\value{
Un objeto de tipo función que formatea los números según el formato especificado.
}
\description{
Esta función asigna un formato específico para los números en función de un parámetro de formato. Los formatos disponibles son:
"coma" (números con comas como separadores de miles), "numero" (números con dos decimales),
"dinero" (números con símbolo de dólar y comas como separadores de miles), "miles" (números en miles con símbolo de dólar),
y "porcentaje" (números expresados como porcentaje con dos decimales).
}
\examples{
# Formato de número con comas
DefinirFormato("coma")(1234567.89)

# Formato de dinero
DefinirFormato("dinero")(1234567.89)

}
