% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Texto.R
\name{EsVacio}
\alias{EsVacio}
\title{Verifica si un valor es NULL, NA o una cadena vacía}
\usage{
EsVacio(x)
}
\arguments{
\item{x}{Un valor a verificar.}
}
\value{
Un valor lógico: TRUE si el valor es NULL, NA o "", de lo contrario FALSE.
}
\description{
Esta función evalúa un valor y determina si es NULL, NA o una cadena vacía ("").
}
\examples{
es_vacio(NULL)    # TRUE
es_vacio(NA)      # TRUE
es_vacio("")      # TRUE
es_vacio("texto") # FALSE
}
