% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Texto.R
\name{EsVacio}
\alias{EsVacio}
\title{Verifica si un valor es NULL, NA o una cadena vacía}
\usage{
EsVacio(x)
}
\arguments{
\item{x}{Valor o vector de valores a verificar.}
}
\value{
Un valor lógico: \code{TRUE} si \code{x} es \code{NULL} o si algún elemento es \code{NA} o una cadena vacía (tras aplicar \code{trimws}); en caso contrario \code{FALSE}.
}
\description{
Esta función evalúa un valor y determina si es \code{NULL}, \code{NA} o una cadena vacía (\code{""}). Se aceptan vectores.
}
\examples{
EsVacio(NULL)              # TRUE
EsVacio(c(NA, ""))        # TRUE
EsVacio(c("texto", NA))   # TRUE
EsVacio(c(" ", "texto")) # TRUE
EsVacio(c("a", "b"))     # FALSE
}
