% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Formatos.R
\name{FormatearNumero}
\alias{FormatearNumero}
\title{Formatear número en HTML}
\usage{
FormatearNumero(
  x,
  formato,
  negrita = TRUE,
  color = "#000000",
  meta = NA,
  prop = TRUE
)
}
\arguments{
\item{x}{Número o vector numérico.}

\item{formato}{Formato numérico (usado en \code{DefinirFormato}).}

\item{negrita}{Lógico, si el número debe mostrarse en negrita (default: TRUE).}

\item{color}{Cadena con el color en formato hexadecimal (default: "#000000").}

\item{meta}{Valor umbral para asignar colores (opcional).}

\item{prop}{Lógico, si los colores son proporcionales respecto a \code{meta}.}
}
\value{
Texto en HTML con estilos aplicados.
}
\description{
Aplica formato numérico y estilos visuales (color, negrita) en HTML.
}
\examples{
FormatearNumero(2500, "dinero", meta = 2000)
FormatearNumero(0.75, "porcentaje", negrita = FALSE, color = "#00FF00")

}
