% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Formatos.R
\name{FormatoHOT}
\alias{FormatoHOT}
\title{Definir formato para Handsontable}
\usage{
FormatoHOT(formato)
}
\arguments{
\item{formato}{Cadena de texto con el formato deseado. Valores permitidos:
"coma", "numero", "dinero", "dolares", "miles", "porcentaje",
"cientifico", "millones", "entero", "tiempo", "kwh" o "log".}
}
\value{
Un string con el formato de Handsontable.
}
\description{
Devuelve un string con el formato numérico para usar en Handsontable.
}
\examples{
FormatoHOT("coma")
FormatoHOT("dinero")

}
