% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Formatos.R
\name{FormatoHOT}
\alias{FormatoHOT}
\title{Definir formato para Handsontable}
\usage{
FormatoHOT(formato)
}
\arguments{
\item{formato}{Cadena de texto que define el formato requerido.}
}
\value{
Un string con el formato de Handsontable.
}
\description{
Devuelve un string con el formato numérico para usar en Handsontable.
Soporta los formatos "coma", "numero", "dinero", "dolares", "miles",
"porcentaje", "cientifico", "millones", "entero", "tiempo", "kwh" y
"log".
}
\examples{
FormatoHOT("coma")
FormatoHOT("dinero")

}
