% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Formatos.R
\name{FormatoJS}
\alias{FormatoJS}
\title{Definir formato para JavaScript}
\usage{
FormatoJS(formato)
}
\arguments{
\item{formato}{Cadena de texto: "coma", "numero", "dinero" o "porcentaje".}
}
\value{
Un string con una función en JavaScript que formatea números.
}
\description{
Genera una función en formato string para aplicar en JavaScript.
}
\examples{
FormatoJS("coma")
FormatoJS("dinero")

}
