% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Formatos.R
\name{FormatoJS}
\alias{FormatoJS}
\title{Definir formato para JavaScript}
\usage{
FormatoJS(formato)
}
\arguments{
\item{formato}{Cadena de texto que define el formato requerido.}
}
\value{
Un string con una función en JavaScript que formatea números.
}
\description{
Genera una función en formato string para aplicar en JavaScript.
Soporta los formatos "coma", "numero", "dinero", "dolares", "miles",
"porcentaje", "cientifico", "millones", "entero", "tiempo", "kwh" y
"log".
}
\examples{
FormatoJS("coma")
FormatoJS("dinero")

}
