% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Formatos.R
\name{FormatoJS}
\alias{FormatoJS}
\title{Definir formato para JavaScript}
\usage{
FormatoJS(formato)
}
\arguments{
\item{formato}{Cadena de texto con el formato deseado. Valores permitidos:
"coma", "numero", "dinero", "dolares", "miles", "porcentaje",
"cientifico", "millones", "entero", "tiempo", "kwh" o "log".}
}
\value{
Un string con una función en JavaScript que formatea números.
}
\description{
Genera una función en formato string para aplicar en JavaScript.
}
\examples{
FormatoJS("coma")
FormatoJS("dinero")

}
