% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Formatos.R
\name{gt_minimal_style}
\alias{gt_minimal_style}
\title{Estilo minimalista para tablas gt}
\usage{
gt_minimal_style(gt_table)
}
\arguments{
\item{gt_table}{Objeto de clase `gt_tbl` generado con el paquete `gt`.}
}
\value{
Un objeto `gt_tbl` con estilos personalizados.
}
\description{
Aplica un diseño limpio y minimalista a un objeto creado con
el paquete **gt**, eliminando bordes y resaltando encabezados con una línea sutil.
}
\examples{
library(gt)
gt(head(mtcars)) \%>\% gt_minimal_style()

}
