% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Outputs.R
\name{ImprimeSankey}
\alias{ImprimeSankey}
\title{Imprime un diagrama de Sankey interactivo}
\usage{
ImprimeSankey(data, vars, fun, var = NULL, colores)
}
\arguments{
\item{data}{Un `data.frame` o `tibble` que contiene los datos de entrada.}

\item{vars}{Un vector de nombres de columnas que se utilizarán para las variables del Sankey.}

\item{fun}{Una cadena que especifica la función de agregación a aplicar. Puede ser "n" para contar
las observaciones o cualquier otra función numérica (e.g., "sum", "mean").}

\item{var}{Un nombre de columna adicional que se usará si `fun` no es "n" (opcional).}

\item{colores}{Un vector de colores para cada variable en `vars`.}
}
\value{
Un objeto `list` que contiene el gráfico de Sankey, los nodos y los enlaces con sus
        respectivos valores y descripciones.
}
\description{
Esta función crea un diagrama de Sankey a partir de los datos proporcionados, con nodos y enlaces
que representan las transiciones entre las categorías de las variables especificadas. Los nodos
están coloreados según las variables y las frecuencias de las transiciones se muestran al pasar el
ratón sobre ellos.
}
\examples{
# Ejemplo de uso
data <- data.frame(Origen = c("TOTAL", "TOTAL", "TOTAL"),
                   SegmentoRacafe = c("A", "B", "C"),
                   Margen = c(100, 200, 300))
result <- ImprimeSankey(data, c("SegmentoRacafe"), "sum", "Margen", c("red", "blue", "green"))
result$plot
}
