% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Outputs.R
\name{ImprimeSankey}
\alias{ImprimeSankey}
\title{Imprime un diagrama de Sankey con datos de agrupación}
\usage{
ImprimeSankey(data, vars, fun, var = NULL, colores)
}
\arguments{
\item{data}{Data frame con los datos de entrada para el diagrama de Sankey.}

\item{vars}{Vector de nombres de columnas a utilizar en el diagrama.}

\item{fun}{Función de agregación a aplicar, como "n" para contar o funciones estadísticas (e.g., "sum", "mean").}

\item{var}{Nombre de la variable sobre la cual se aplica `fun` (opcional).}

\item{colores}{Vector de colores a aplicar a cada variable especificada en `vars`.}
}
\value{
Una lista con el gráfico Sankey (`plot`) y los datos de nodos y enlaces (`nodos`, `arcos`).
}
\description{
Esta función crea un diagrama de Sankey con la biblioteca `plotly`, basado en las agrupaciones especificadas
en los datos de entrada y permite visualizar relaciones entre variables.
}
\examples{
ImprimeSankey(data = df, vars = c("Var1", "Var2"), fun = "n", colores = c("blue", "green"))

}
