% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Outputs.R
\name{ImprimeSankey}
\alias{ImprimeSankey}
\title{Genera un diagrama de Sankey basado en un conjunto de datos}
\usage{
ImprimeSankey(data, vars, fun, var = NULL, colores)
}
\arguments{
\item{data}{Data frame que contiene los datos a graficar.}

\item{vars}{Vector de nombres de variables que serán utilizadas en el diagrama de Sankey.}

\item{fun}{Función de resumen a aplicar (por ejemplo, "n" para el conteo de filas).}

\item{var}{Nombre de la variable que se utilizará en la función de resumen (solo necesario si `fun` no es "n").}

\item{colores}{Vector de colores que corresponde a cada una de las variables especificadas en `vars`.}
}
\value{
Una lista con dos elementos: el gráfico de Sankey (`plot`) y los data frames de nodos y arcos.
}
\description{
La función `ImprimeSankey` crea un diagrama de Sankey interactivo a partir de un conjunto de datos, utilizando variables y colores definidos por el usuario.
}
\examples{
# Ejemplo de uso
ImprimeSankey(data = df, vars = c("variable1", "variable2"), fun = "mean", var = "variable3", colores = c("blue", "green"))
}
