% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Html.R
\name{Obligatorio}
\alias{Obligatorio}
\title{Obligatorio: Campo obligatorio con un asterisco rojo}
\usage{
Obligatorio(s)
}
\arguments{
\item{s}{Texto que se mostrará como etiqueta del campo obligatorio.}
}
\value{
Un objeto HTML que contiene el texto de la etiqueta con un asterisco rojo que indica que el campo es obligatorio.
}
\description{
Esta función genera un componente visual de campo obligatorio con un asterisco rojo al lado de la etiqueta.
}
\examples{
Obligatorio("Nombre")  # Añade un asterisco rojo a la etiqueta "Nombre" para indicar que es obligatorio.
}
