% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datos.R
\name{ObtenerTokenAcceso}
\alias{ObtenerTokenAcceso}
\title{ObtenerTokenAcceso}
\usage{
ObtenerTokenAcceso()
}
\value{
Cadena con el token de acceso.
}
\description{
Obtiene un token de acceso (client_credentials) para Microsoft Graph API usando las variables de ambiente `MS_CLIENT_ID`, `MS_CLIENT_SECRET` y `MS_TENANT_ID`.
}
\examples{
# token <- ObtenerTokenAcceso()
}
