% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Numericos.R
\name{Variacion}
\alias{Variacion}
\title{Calcula la variación relativa entre dos valores}
\usage{
Variacion(ini, fin)
}
\arguments{
\item{ini}{Vector numérico con los valores iniciales.}

\item{fin}{Vector numérico con los valores finales.}
}
\value{
Un vector numérico con la variación relativa entre `ini` y `fin`.
  - Retorna `NA` si alguno de los valores es `NA`.
  - Retorna `0` si ambos valores son `0`.
  - Retorna `sign(y)` si `ini` es `0` y `fin` no es `0`.
  - En otros casos, retorna `(y - x) / abs(x)`.
}
\description{
Esta función calcula la variación relativa entre dos vectores de valores iniciales y finales.
Si el valor inicial es cero, se define una variación especial basada en el signo del valor final.
}
\examples{
Variacion(c(10, 0, NA, 5), c(15, 5, 20, 10))
# Retorna: c(0.5, 1, NA, 1)
}
